7335 Extract FieldValueSerDe
