+++ java/core/src/main/java/sleeper/core/row/serialiser/RowJsonSerDe.java
@@ -17,33 +17,17 @@ package sleeper.core.row.serialiser;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.JsonArray;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonNull;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
-import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializationContext;
 import com.google.gson.JsonSerializer;
 
 import sleeper.core.row.Row;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
-import sleeper.core.schema.type.ByteArrayType;
-import sleeper.core.schema.type.IntType;
-import sleeper.core.schema.type.ListType;
-import sleeper.core.schema.type.LongType;
-import sleeper.core.schema.type.MapType;
-import sleeper.core.schema.type.PrimitiveType;
-import sleeper.core.schema.type.StringType;
-
-import java.util.ArrayList;
-import java.util.Base64;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 /**
  * Serialises and deserialises a row to and from a JSON string.
@@ -112,7 +96,8 @@ public class RowJsonSerDe {
         public JsonElement serialize(Row row, java.lang.reflect.Type typeOfSrc, JsonSerializationContext context) {
             JsonObject json = new JsonObject();
             for (Field field : schema.getAllFields()) {
-                addFieldToJsonObject(field, row.get(field.getName()), json);
+                Object value = row.get(field.getName());
+                json.add(field.getName(), FieldValueSerDe.toJsonElement(field, value));
             }
             return json;
         }
@@ -122,177 +107,16 @@ public class RowJsonSerDe {
             if (!jsonElement.isJsonObject()) {
                 throw new JsonParseException("Expected JsonObject, got " + jsonElement);
             }
+            JsonObject jsonObject = jsonElement.getAsJsonObject();
             Row row = new Row();
             for (Field field : schema.getAllFields()) {
-                getFieldFromJsonObject(field, jsonElement.getAsJsonObject(), row);
+                Object value = FieldValueSerDe.fromJsonElement(field, jsonObject.get(field.getName()));
+                if (value == null && !field.isNullable()) {
+                    continue;
+                }
+                row.put(field.getName(), value);
             }
             return row;
         }
-    }
-
-    private static void addFieldToJsonObject(Field field, Object fieldValue, JsonObject json) {
-        json.add(field.getName(), toJsonElement(field, fieldValue));
-    }
-
-    public static JsonElement toJsonElement(Field field, Object fieldValue) {
-        if (fieldValue == null) {
-            return JsonNull.INSTANCE;
-        } else if (field.getType() instanceof IntType) {
-            return new JsonPrimitive((Integer) fieldValue);
-        } else if (field.getType() instanceof LongType) {
-            return new JsonPrimitive((Long) fieldValue);
-        } else if (field.getType() instanceof StringType) {
-            return new JsonPrimitive((String) fieldValue);
-        } else if (field.getType() instanceof ByteArrayType) {
-            byte[] bytes = (byte[]) fieldValue;
-            String base64encodedBytes = Base64.getEncoder().encodeToString(bytes);
-            return new JsonPrimitive(base64encodedBytes);
-        } else if (field.getType() instanceof ListType) {
-            return listToJsonElement(field, (List<Object>) fieldValue);
-        } else if (field.getType() instanceof MapType) {
-            return mapToJsonElement(field, (Map<Object, Object>) fieldValue);
-        } else {
-            throw new IllegalArgumentException("Unknown type " + field.getType());
-        }
-    }
-
-    private static JsonElement listToJsonElement(Field field, List<Object> fieldValue) {
-        PrimitiveType elementType = ((ListType) field.getType()).getElementType();
-        JsonArray array = new JsonArray();
-        if (elementType instanceof IntType) {
-            for (Object o : fieldValue) {
-                array.add((Integer) o);
-            }
-        } else if (elementType instanceof LongType) {
-            for (Object o : fieldValue) {
-                array.add((Long) o);
-            }
-        } else if (elementType instanceof StringType) {
-            for (Object o : fieldValue) {
-                array.add((String) o);
-            }
-        } else if (elementType instanceof ByteArrayType) {
-            for (Object o : fieldValue) {
-                if (null != o) {
-                    array.add(Base64.getEncoder().encodeToString((byte[]) o));
-                } else {
-                    array.add(JsonNull.INSTANCE);
-                }
-            }
-        } else {
-            throw new IllegalArgumentException("Unknown type " + field.getType());
-        }
-        return array;
-    }
-
-    private static JsonElement mapToJsonElement(Field field, Map<Object, Object> fieldValue) {
-        PrimitiveType keyType = ((MapType) field.getType()).getKeyType();
-        PrimitiveType valueType = ((MapType) field.getType()).getValueType();
-
-        JsonObject map = new JsonObject();
-        for (Map.Entry<Object, Object> entry : fieldValue.entrySet()) {
-            String key;
-            if (keyType instanceof IntType || keyType instanceof LongType || keyType instanceof StringType) {
-                key = entry.getKey().toString();
-            } else if (keyType instanceof ByteArrayType) {
-                byte[] bytes = (byte[]) entry.getKey();
-                key = Base64.getEncoder().encodeToString(bytes);
-            } else {
-                throw new IllegalArgumentException("Unknown type " + field.getType());
-            }
-            if (valueType instanceof IntType || valueType instanceof LongType) {
-                map.addProperty(key, (Number) entry.getValue());
-            } else if (valueType instanceof StringType) {
-                map.addProperty(key, (String) entry.getValue());
-            } else {
-                throw new IllegalArgumentException("Unknown type " + field.getType());
-            }
-        }
-        return map;
-    }
-
-    private static void getFieldFromJsonObject(Field field, JsonObject json, Row row) {
-        JsonElement element = json.get(field.getName());
-        if (element == null || element.isJsonNull()) {
-            if (field.isNullable()) {
-                row.put(field.getName(), null);
-            }
-            return;
-        }
-        if (field.getType() instanceof IntType) {
-            row.put(field.getName(), element.getAsInt());
-        } else if (field.getType() instanceof LongType) {
-            row.put(field.getName(), element.getAsLong());
-        } else if (field.getType() instanceof StringType) {
-            row.put(field.getName(), element.getAsString());
-        } else if (field.getType() instanceof ByteArrayType) {
-            row.put(field.getName(), Base64.getDecoder().decode(element.getAsString()));
-        } else if (field.getType() instanceof ListType) {
-            getListFromJsonObject(field, json, row);
-        } else if (field.getType() instanceof MapType) {
-            getMapFromJsonObject(field, json, row);
-        } else {
-            throw new IllegalArgumentException("Unknown type " + field.getType());
-        }
-    }
-
-    private static void getListFromJsonObject(Field field, JsonObject json, Row row) {
-        PrimitiveType elementType = ((ListType) field.getType()).getElementType();
-        JsonArray array = json.get(field.getName()).getAsJsonArray();
-        List<Object> list = new ArrayList<>();
-        for (int i = 0; i < array.size(); i++) {
-            if (elementType instanceof IntType) {
-                list.add(array.get(i).getAsInt());
-            } else if (elementType instanceof LongType) {
-                list.add(array.get(i).getAsLong());
-            } else if (elementType instanceof StringType) {
-                list.add(array.get(i).getAsString());
-            } else if (elementType instanceof ByteArrayType) {
-                String encodedByteArray = array.get(i).getAsString();
-                list.add(Base64.getDecoder().decode(encodedByteArray));
-            } else {
-                throw new IllegalArgumentException("Unknown type " + elementType);
-            }
-        }
-        row.put(field.getName(), list);
-    }
-
-    private static void getMapFromJsonObject(Field field, JsonObject json, Row row) {
-        PrimitiveType keyType = ((MapType) field.getType()).getKeyType();
-        PrimitiveType valueType = ((MapType) field.getType()).getValueType();
-
-        JsonObject map = json.getAsJsonObject(field.getName());
-        Map<Object, Object> deserialisedMap = new HashMap<>();
-        for (Map.Entry<String, JsonElement> entry : map.entrySet()) {
-            String keyString = entry.getKey();
-            Object key;
-            if (keyType instanceof IntType) {
-                key = Integer.parseInt(keyString);
-            } else if (keyType instanceof LongType) {
-                key = Long.parseLong(keyString);
-            } else if (keyType instanceof StringType) {
-                key = keyString;
-            } else if (keyType instanceof ByteArrayType) {
-                key = Base64.getDecoder().decode(keyString);
-            } else {
-                throw new IllegalArgumentException("Unknown type " + keyType);
-            }
-            JsonElement valueElement = entry.getValue();
-            Object value;
-            if (valueType instanceof IntType) {
-                value = valueElement.getAsInt();
-            } else if (valueType instanceof LongType) {
-                value = valueElement.getAsLong();
-            } else if (valueType instanceof StringType) {
-                value = valueElement.getAsString();
-            } else if (valueType instanceof ByteArrayType) {
-                String encodedByteArray = valueElement.getAsString();
-                value = Base64.getDecoder().decode(encodedByteArray);
-            } else {
-                throw new IllegalArgumentException("Unknown type " + keyType);
-            }
-            deserialisedMap.put(key, value);
-        }
-        row.put(field.getName(), deserialisedMap);
     }
 }
